% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{jpplantnames-package}
\alias{jpplantnames}
\alias{jpplantnames-package}
\title{Unofficial Japanese Plant Name Lookup Tools}
\description{
Fetches and caches the Vascular Plant Japanese Name Checklist ver. 1.10 and
provides helpers for looking up scientific plant names from Japanese names.
}
\details{
This package is not affiliated with or endorsed by JBIF, YList, or the
checklist authors. Checklist data is not included in the package and is not
covered by the package license.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/maple60/jpplantnames}
  \item \url{https://maple60.github.io/jpplantnames/}
  \item Report bugs at \url{https://github.com/maple60/jpplantnames/issues}
}

}
\author{
\strong{Maintainer}: Kaede Konrai \email{konrai630@gmail.com}

Authors:
\itemize{
  \item Kaede Konrai \email{konrai630@gmail.com}
}

}
\keyword{internal}
